requirements.txt
torch~=1.9.0
transformers~=4.12.0

fastapi
uvicorn[standard]

python-dotenv
pyGithub
requests
